Make message-field into a component
